Push new line when finding newline character in text mesh generation
